PR B: clear 2nd-review cosmetic nits
- VerseReferencePill: inline the trivial `pointSize` alias (it became a
  pure passthrough to `basePoint` once the fontScale multiply was dropped);
  the three call sites read `basePoint` directly.
- MessageListSnapshotTests: drop the redundant `CGFloat(fontScale)` cast —
  the `SuperTypography.make` bridge takes a `Double`, so `fontScale` passes
  straight through.

+++ Packages/Chat/Tests/ChatTests/UI/Snapshots/MessageListSnapshotTests.swift
@@ -705,7 +705,7 @@ struct MessageListSnapshotTests {
         let view = MessageList(items: items, verbosity: .verbose)
             .superTheme(.make(theme))
             .chatAppearance(ChatAppearance(fontScale: fontScale))
-            .superTypography(.make(.serif, fontScale: CGFloat(fontScale)))
+            .superTypography(.make(.serif, fontScale: fontScale))
             .frame(width: 402, height: 700)
         recordOrCompare(view: view, name: name, function: function)
     }
